requirements.txt
boto3==1.37.6
botocore==1.37.6
jmespath==1.0.1
loguru==0.7.3
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
